# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family VPN — local authenticated stand",
    "version": "0.13.0",
    "description": "Implemented local demo, HTTPS user authentication and separate password/TOTP admin authentication. Read-only management views. No production bind or real VPN delivery."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Portal demo"
    },
    {
      "url": "http://127.0.0.1:8081",
      "description": "Admin demo"
    },
    {
      "url": "https://127.0.0.1:8443",
      "description": "Local user auth"
    },
    {
      "url": "https://127.0.0.1:9443",
      "description": "Local admin TOTP"
    }
  ],
  "paths": {
    "/api/v1/me": {
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "devices",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Device"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      },
      "post": {
        "summary": "Local user device intent: /devices",
        "description": "Authenticated owner only. Mutations require Origin/CSRF and JSON. Request creates two pending profile slots without keys or peer provisioning. Cancellation only before any configured profile/subscription; it is not a network revoke. request_id is owner-scoped idempotency, same payload required; expected_revision protects updates.",
        "responses": {
          "200": {
            "description": "Current result (including idempotent replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Login required"
          },
          "403": {
            "description": "Origin/CSRF rejected"
          },
          "404": {
            "description": "Not found or not owned"
          },
          "409": {
            "description": "Limit, revision, idempotency payload or state conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}": {
      "get": {
        "operationId": "devices_by_id",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "status",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Health"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
    "/api/v1/instructions": {
      "get": {
        "operationId": "instructions",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Instruction"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
    "/api/v1/admin/overview": {
      "get": {
        "operationId": "admin_overview",
        "responses": {
          "200": {
            "description": "Read-only demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "default": {
            "description": "API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Available ONLY on the admin listener :8081. Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
    "/api/v1/profiles/{id}/download": {
      "get": {
        "operationId": "downloadDemoProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "txt"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NONFUNCTIONAL demo note; no VPN keys",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Not found/not ready/unsupported format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": " Local-auth requires user session; demo remains explicit fixed fixtures. Management views require the separate password/TOTP admin listener."
      }
    },
...
  },
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "required": [
          "id",
          "protocol",
          "state",
          "format"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "awg",
              "reality"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "ready",
              "pending",
              "revoking",
              "revoked",
              "error"
            ]
          },
          "format": {
            "type": "string",
            "enum": [
              "txt",
              "",
              "vless-reality-uri",
              "awg-3.1-conf"
            ],
            "description": "Empty means unselected; txt is demo only; vless-reality-uri and awg-3.1-conf are imported client data that remain pending until installation verification."
          }
        }
      },
      "Device": {
        "type": "object",
        "required": [
          "id",
          "name",
          "os",
          "state",
          "profiles",
          "revision"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "os": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "profiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Profile"
            }
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "Health": {
        "type": "object",
        "required": [
          "component",
          "status",
          "reason",
          "measured_at",
          "expires_at"
        ],
        "properties": {
          "component": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "unavailable",
              "unknown"
            ]
          },
          "reason": {
            "type": "string"
          },
          "measured_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Instruction": {
        "type": "object",
        "required": [
          "id",
          "os",
          "title",
          "verified",
          "steps",
          "offline_available"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "os": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "kind": {
            "type": "string",
            "enum": [
              "portal",
              "vpn"
            ]
          },
          "content_version": {
            "type": "integer",
            "minimum": 1
          },
          "verified_at": {
            "type": "string",
            "format": "date"
          },
          "verification_scope": {
            "type": "string",
            "enum": [
              "none",
              "portal",
              "client"
            ]
          },
          "app_id": {
            "type": "string"
          },
          "app_version": {
            "type": "string"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "awg",
              "reality"
            ]
          },
          "format": {
            "type": "string"
          },
          "offline_available": {
            "type": "boolean"
          }
        }
      },
      "Me": {
        "type": "object",
        "required": [
          "id",
          "display_name",
          "role",
          "mode",
          "storage"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "local-auth",
              "local-admin"
            ]
          },
          "storage": {
            "type": "string",
            "enum": [
              "sqlite",
              "memory-fixture"
            ]
          },
          "csrf_token": {
            "type": "string",
            "description": "Only local-auth; held in JS memory, never localStorage."
          }
        }
      },
      "Overview": {
        "type": "object",
        "required": [
          "mode",
          "total_devices",
          "mutations_enabled",
          "message",
          "components"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "local-admin"
            ]
          },
          "total_devices": {
            "type": "integer"
          },
          "mutations_enabled": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "DeviceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "request_id",
          "name",
          "os"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 16,
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "os": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "windows",
              "macos",
              "linux",
              "other"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-fvpn_session"
      },
      "preauthCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-fvpn_preauth"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "adminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-fvpn_admin_session"
      }
    }
  }
}